Move syncing code to the sync_strategies folder
